+++ .php-cs-fixer.dist.php
@@ -3,8 +3,17 @@
 $finder = (new PhpCsFixer\Finder())
     ->in(__DIR__.'/src')
     ->in(__DIR__.'/tests')
+    ->in(__DIR__.'/tools')
     ->exclude('Fixtures/RuntimeApplication')
-    ->append([__FILE__, __DIR__.'/bin/symfony-lsp'])
+    ->append([
+        __FILE__,
+        __DIR__.'/bin/symfony-lsp',
+        __DIR__.'/tools/benchmark-server',
+        __DIR__.'/tools/dogfood-server',
+        __DIR__.'/tools/dogfood-symfonycorp',
+        __DIR__.'/tools/release',
+        __DIR__.'/tools/smoke-test-server',
+    ])
 ;
 
 return (new PhpCsFixer\Config())